requirements.txt
asn1crypto==0.24.0
astroid==2.3.2
cachetools==3.1.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-speech==1.2.0
google-cloud-translate==1.7.0
googleapis-common-protos==1.6.0
grpcio==1.24.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pkg-resources==0.0.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pylint==2.4.3
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.3
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
